requirements.txt
mindspore==2.5.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.65.0 